updated code, added graph for dashboards

+++ stop_ssds.sh
@@ -2,6 +2,13 @@
 
 # Stop the firewall daemons
 
+TMP=`ps -ef`
+TMP=`echo "$TMP" | fgrep "wash_blocked.sp"`
+if [ -n "$TMP" ] ; then
+   echo "wash_blocked is still running"
+   exit 192
+fi
+
 if [ -f "run/http_daemon.pid" ] ; then
    kill `cat run/http_daemon.pid`
 fi
